refactor(loop-support): extract capability surface snapshot model (#3955)
