+++ crates/ironclaw_loop_support/src/capability_port.rs
@@ -6,7 +6,7 @@ use std::{
 use async_trait::async_trait;
 use ironclaw_host_api::{
     CapabilityId, CapabilitySet, CorrelationId, EffectKind, ExecutionContext, ExtensionId,
-    InvocationId, MountView, Principal, ResourceEstimate, RuntimeKind, sha256_digest_token,
+    InvocationId, MountView, Principal, sha256_digest_token,
 };
 use ironclaw_host_runtime::{
     HostRuntime, HostRuntimeError, IdempotencyKey, RuntimeBlockedReason, RuntimeCapabilityOutcome,
@@ -27,7 +27,12 @@ use ironclaw_turns::{
 };
 use tokio::sync::Notify;
 
-use crate::capability_info::{self, CapabilityInfoEntry};
+mod surface_snapshot;
+
+use self::surface_snapshot::{
+    RuntimeSurfaceCapabilitySnapshot, SurfaceCapabilitySnapshot, SurfaceSnapshot,
+    SyntheticSurfaceCapabilitySnapshot,
+};
 
 const PROVIDER_TOOL_NAME_MAX_BYTES: usize = 64;
 const PROVIDER_TOOL_NAME_DIGEST_BYTES: usize = 32;
@@ -189,252 +194,6 @@ impl HostRuntimeLoopCapabilityPortFactory {
         .with_execution_mounts(self.execution_mounts.clone())
         .with_capability_execution_mounts(self.capability_execution_mounts.clone())
     }
-}
-
-#[derive(Clone)]
-struct RuntimeSurfaceCapabilitySnapshot {
-    provider: ExtensionId,
-    runtime: RuntimeKind,
-    estimate: ResourceEstimate,
-    safe_description: String,
-    parameters_schema: serde_json::Value,
-    effects: Vec<EffectKind>,
-    provider_tool_name: String,
-}
-
-#[derive(Clone)]
-struct SyntheticSurfaceCapabilitySnapshot {
-    provider_tool_name: String,
-    kind: SyntheticCapabilityKind,
-}
-
-#[derive(Clone, Copy)]
-enum SyntheticCapabilityKind {
-    CapabilityInfo,
-}
-
-#[derive(Clone)]
-enum SurfaceCapabilitySnapshot {
-    Runtime(Box<RuntimeSurfaceCapabilitySnapshot>),
-    Synthetic(SyntheticSurfaceCapabilitySnapshot),
-}
-
-#[derive(Clone, Default)]
-struct SurfaceSnapshot {
-    capabilities: HashMap<CapabilityId, SurfaceCapabilitySnapshot>,
-    provider_names: HashMap<String, CapabilityId>,
-}
-
-impl SurfaceSnapshot {
-    fn insert_synthetic_capabilities(&mut self) -> Result<(), AgentLoopHostError> {
-        let capability_id = capability_info::capability_id()?;
-        self.provider_names.insert(
-            capability_info::TOOL_NAME.to_string(),
-            capability_id.clone(),
-        );
-        self.capabilities.insert(
-            capability_id,
-            SurfaceCapabilitySnapshot::Synthetic(SyntheticSurfaceCapabilitySnapshot {
-                provider_tool_name: capability_info::TOOL_NAME.to_string(),
-                kind: SyntheticCapabilityKind::CapabilityInfo,
-            }),
-        );
-        Ok(())
-    }
-
-    fn capability_info(&self, requested: &str) -> Option<CapabilityInfoEntry<'_>> {
-        if let Some(capability_id) = self.provider_names.get(requested)
-            && let Some(capability) = self
-                .capabilities
-                .get(capability_id)
-                .and_then(SurfaceCapabilitySnapshot::as_runtime)
-        {
-            return Some(capability.capability_info(capability_id));
-        }
-        let requested_id = CapabilityId::new(requested).ok()?;
-        self.capabilities
-            .get_key_value(&requested_id)
-            .and_then(|(capability_id, capability)| {
-                capability
-                    .as_runtime()
-                    .map(|capability| (capability_id, capability))
-            })
-            .map(|(capability_id, capability)| capability.capability_info(capability_id))
-    }
-
-    fn provider_capability(
-        &self,
-        provider_tool_name: &str,
-    ) -> Result<(&CapabilityId, &SurfaceCapabilitySnapshot), AgentLoopHostError> {
-        let Some(capability_id) = self.provider_names.get(provider_tool_name) else {
-            return Err(AgentLoopHostError::new(
-                AgentLoopHostErrorKind::InvalidInvocation,
-                "provider tool call is outside the visible capability surface",
-            ));
-        };
-        let Some(capability) = self.capabilities.get(capability_id) else {
-            return Err(AgentLoopHostError::new(
-                AgentLoopHostErrorKind::StaleSurface,
-                "capability surface snapshot is missing provider metadata",
-            ));
-        };
-        Ok((capability_id, capability))
-    }
-}
-
-impl RuntimeSurfaceCapabilitySnapshot {
-    fn capability_info<'a>(&'a self, capability_id: &'a CapabilityId) -> CapabilityInfoEntry<'a> {
-        CapabilityInfoEntry {
-            capability_id,
-            provider_tool_name: &self.provider_tool_name,
-            safe_description: &self.safe_description,
-            parameters_schema: &self.parameters_schema,
-            runtime: self.runtime,
-            effects: &self.effects,
-        }
-    }
-}
-
-impl SurfaceCapabilitySnapshot {
-    fn as_runtime(&self) -> Option<&RuntimeSurfaceCapabilitySnapshot> {
-        match self {
-            Self::Runtime(capability) => Some(capability.as_ref()),
-            Self::Synthetic(_) => None,
-        }
-    }
-
-    fn tool_definition(
-        &self,
-        capability_id: &CapabilityId,
-    ) -> Result<Option<ProviderToolDefinition>, AgentLoopHostError> {
-        match self {
-            Self::Runtime(capability) => {
-                if !provider_schema_is_usable(&capability.parameters_schema) {
-                    tracing::debug!(
-                        capability_id = capability_id.as_str(),
-                        "capability omitted from provider tool definitions because its parameter schema is not provider-usable"
-                    );
-                    return Ok(None);
-                }
-                Ok(Some(ProviderToolDefinition {
-                    capability_id: capability_id.clone(),
-                    name: capability.provider_tool_name.clone(),
-                    description: capability.safe_description.clone(),
-                    parameters: capability.parameters_schema.clone(),
-                }))
-            }
-            Self::Synthetic(capability) => capability.tool_definition(capability_id).map(Some),
-        }
-    }
-
-    fn prepare_provider_tool_call(
-        &self,
-        capability_id: &CapabilityId,
-        snapshot: &SurfaceSnapshot,
-        tool_call: &ProviderToolCall,
-    ) -> Result<PreparedSurfaceCapabilityCall, AgentLoopHostError> {
-        match self {
-            Self::Runtime(capability) => {
-                capability.prepare_provider_tool_call(capability_id, tool_call)
-            }
-            Self::Synthetic(capability) => {
-                capability.prepare_provider_tool_call(capability_id, snapshot, tool_call)
-            }
-        }
-    }
-}
-
-impl RuntimeSurfaceCapabilitySnapshot {
-    fn prepare_provider_tool_call(
-        &self,
-        capability_id: &CapabilityId,
-        tool_call: &ProviderToolCall,
-    ) -> Result<PreparedSurfaceCapabilityCall, AgentLoopHostError> {
-        if !provider_schema_is_usable(&self.parameters_schema) {
-            return Err(AgentLoopHostError::new(
-                AgentLoopHostErrorKind::InvalidInvocation,
-                "provider tool call was not advertised to the model",
-            ));
-        }
-        let normalized_arguments = normalize_provider_arguments(
-            &tool_call.arguments,
-            &self.parameters_schema,
-            "provider arguments",
-        )?;
-        validate_provider_arguments(&normalized_arguments)?;
-        Ok(PreparedSurfaceCapabilityCall {
-            capability_id: capability_id.clone(),
-            normalized_arguments,
-            effective_capability_ids: vec![capability_id.clone()],
-        })
-    }
-}
-
-impl SyntheticSurfaceCapabilitySnapshot {
-    fn tool_definition(
-        &self,
-        capability_id: &CapabilityId,
-    ) -> Result<ProviderToolDefinition, AgentLoopHostError> {
-        match self.kind {
-            SyntheticCapabilityKind::CapabilityInfo => {
-                debug_assert!(capability_info::is_capability_id(capability_id));
-                let mut definition = capability_info::tool_definition()?;
-                definition.name = self.provider_tool_name.clone();
-                Ok(definition)
-            }
-        }
-    }
-
-    fn prepare_provider_tool_call(
-        &self,
-        capability_id: &CapabilityId,
-        snapshot: &SurfaceSnapshot,
-        tool_call: &ProviderToolCall,
-    ) -> Result<PreparedSurfaceCapabilityCall, AgentLoopHostError> {
-        match self.kind {
-            SyntheticCapabilityKind::CapabilityInfo => {
-                let normalized_arguments = normalize_provider_arguments(
-                    &tool_call.arguments,
-                    &capability_info::schema(),
-                    "provider arguments",
-                )?;
-                validate_provider_arguments(&normalized_arguments)?;
-                let request = capability_info::CapabilityInfoRequest::parse(&normalized_arguments)?;
-                let target = snapshot
-                    .capability_info(request.requested_name())
-                    .ok_or_else(|| {
-                        AgentLoopHostError::new(
-                            AgentLoopHostErrorKind::InvalidInvocation,
-                            "capability_info target is not on the visible surface",
-                        )
-                    })?;
-                Ok(PreparedSurfaceCapabilityCall {
-                    capability_id: capability_id.clone(),
-                    normalized_arguments,
-                    effective_capability_ids: vec![
-                        capability_id.clone(),
-                        target.capability_id.clone(),
-                    ],
-                })
-            }
-        }
-    }
-
-    fn output<'a>(
-        &self,
-        input: &serde_json::Value,
-        resolve: impl FnOnce(&str) -> Option<CapabilityInfoEntry<'a>>,
-    ) -> Result<serde_json::Value, AgentLoopHostError> {
-        match self.kind {
-            SyntheticCapabilityKind::CapabilityInfo => capability_info::output(input, resolve),
-        }
-    }
-}
-
-struct PreparedSurfaceCapabilityCall {
-    capability_id: CapabilityId,
-    normalized_arguments: serde_json::Value,
-    effective_capability_ids: Vec<CapabilityId>,
 }
 
 struct PreparedProviderToolCall {
@@ -1000,8 +759,7 @@ impl LoopCapabilityPort for HostRuntimeLoopCapabilityPort {
             .await
             .map_err(host_runtime_error)?;
         let version = loop_surface_version(runtime_surface.version.as_str())?;
-        let mut snapshot = SurfaceSnapshot::default();
-        snapshot.insert_synthetic_capabilities()?;
+        let mut snapshot = SurfaceSnapshot::with_synthetic_capabilities()?;
         let descriptors = runtime_surface
             .capabilities
             .into_iter()
@@ -2090,7 +1848,7 @@ mod tests {
     use ironclaw_host_api::{
         AgentId, CapabilityDescriptor, CapabilityGrant, CapabilityGrantId, GrantConstraints,
         MountAlias, MountGrant, MountPermissions, NetworkPolicy, PermissionMode, ProjectId,
-        ResourceUsage, TenantId, TrustClass, UserId, VirtualPath,
+        ResourceEstimate, ResourceUsage, RuntimeKind, TenantId, TrustClass, UserId, VirtualPath,
     };
     use ironclaw_host_runtime::{
         CancelRuntimeWorkOutcome, CancelRuntimeWorkRequest, CapabilitySurfaceVersion,
@@ -2103,6 +1861,8 @@ mod tests {
         InMemoryRunProfileResolver, LoopDriverId, RunProfileResolutionRequest, RunProfileResolver,
         TurnId, TurnRunId, TurnScope,
     };
+
+    use crate::capability_info;
 
     #[test]
     fn concurrency_hint_treats_empty_effects_as_exclusive() {
